Rename clean_occurrences to wrangle_occurrences
