+++ lmpy/tools/__init__.py
@@ -1,7 +1,6 @@
 """Tools module."""
 from . import build_shapegrid
 from . import calculate_pam_stats
-from . import clean_occurrences
 from . import convert_csv_to_lmm
 from . import convert_lmm_to_csv
 from . import create_rare_species_model
@@ -11,12 +10,13 @@ from . import mcpa_run
 from . import randomize_pam
 from . import split_occurrence_data
 from . import wrangle_matrix
+from . import wrangle_occurrences
+from . import wrangle_species_list
 from . import wrangle_tree
 
 __all__ = []
 __all__.extend(build_shapegrid.__all__)
 __all__.extend(calculate_pam_stats.__all__)
-__all__.extend(clean_occurrences.__all__)
 __all__.extend(convert_csv_to_lmm.__all__)
 __all__.extend(convert_lmm_to_csv.__all__)
 __all__.extend(create_rare_species_model.__all__)
@@ -26,4 +26,6 @@ __all__.extend(mcpa_run.__all__)
 __all__.extend(randomize_pam.__all__)
 __all__.extend(split_occurrence_data.__all__)
 __all__.extend(wrangle_matrix.__all__)
+__all__.extend(wrangle_occurrences.__all__)
+__all__.extend(wrangle_species_list.__all__)
 __all__.extend(wrangle_tree.__all__)
