+++ setup.cfg
@@ -34,7 +34,6 @@ where = lmpy
 console_scripts =
     build_shapegrid = lmpy.tools.build_shapegrid:cli
     calculate_pam_stats = lmpy.tools.calculate_pam_stats:cli
-    clean_occurrences = lmpy.tools.clean_occurrences:cli
     convert_csv_to_lmm = lmpy.tools.convert_csv_to_lmm:cli
     convert_lmpy_to_csv = lmpy.tools.convert_lmpy_to_csv:cli
     convert_lmpy_to_geojson = lmpy.tools.convert_lmpy_to_geojson:cli
@@ -45,6 +44,8 @@ console_scripts =
     randomize_pam = lmpy.tools.randomize_pam:cli
     split_occurrence_data = lmpy.tools.split_occurrence_data:cli
     wrangle_matrix = lmpy.tools.wrangle_matrix:cli
+    wrangle_occurrences = lmpy.tools.wrangle_occurrences:cli
+    wrangle_species_list = lmpy.tools.wrangle_species_list:cli
     wrangle_tree = lmpy.tools.wrangle_tree:cli
 
 [versioneer]
